requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.39.0

# ASR dependencies
funasr>=1.0.0
modelscope>=1.8.0
onnxruntime>=1.15.0

# TTS dependencies
TTS>=0.15.0
# coqui-ai-tts>=0.15.0

# Audio processing
pyaudio>=0.2.11
sounddevice>=0.4.6
librosa>=0.10.0
scipy>=1.10.0
numpy>=1.24.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0

# Optional for better performance
# torch-audio>=2.0.0
